requirements.txt
Django==5.0.7
mysqlclient
pillow
crispy-bootstrap5
django-crispy-forms
djangorestframework
requests


